Split up base functionality from iter.lua into base.lua.
Adapt Makefile and gslext.lua as well.
